Fixed #99, probably fixed #96 (sorry @Giappo)

+++ R/est_evidences.R
@@ -54,9 +54,15 @@ est_evidences <- function(
   )
   # Delete files
   for (experiment in experiments) {
-    file.remove(experiment$beast2_options$output_log_filename)
-    file.remove(experiment$beast2_options$output_state_filename)
-    file.remove(experiment$beast2_options$output_trees_filenames)
+    if (file.exists(experiment$beast2_options$output_log_filename)) {
+      file.remove(experiment$beast2_options$output_log_filename)
+    }
+    if (file.exists(experiment$beast2_options$output_state_filename)) {
+      file.remove(experiment$beast2_options$output_state_filename)
+    }
+    if (file.exists(experiment$beast2_options$output_trees_filenames)) {
+      file.remove(experiment$beast2_options$output_trees_filenames)
+    }
   }
